+++ src/main/java/webserver/Body.java
@@ -1,9 +1,10 @@
-package utils;
+package webserver;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Optional;
+import utils.IOUtils;
 
 public class Body {
