+++ src/main/java/webserver/HttpMethod.java
@@ -1,4 +1,4 @@
-package utils;
+package webserver;
 
 import java.util.NoSuchElementException;
 import java.util.stream.Stream;
