+++ src/main/java/webserver/Request.java
@@ -1,4 +1,4 @@
-package utils;
+package webserver;
 
 import java.io.BufferedReader;
 import java.io.IOException;
